requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.23
alembic>=1.12.1
asyncpg>=0.28.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.1.0.post1

# Data Processing
pandas>=2.1.2
numpy>=1.26.1

# Configuration
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
watchfiles>=0.21.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.1

# Optional Dependencies
plotly>=5.17.0